Revert "Ensure glyph position is non-negative"
This reverts commit fafa54b5ba6c33bb5d45ac5e12042f28adec2c54.

+++ src/text.rs
@@ -98,7 +98,7 @@ impl Rasterizer {
 
         let mut glyphs_iter = glyphs.iter().peekable();
         while let Some(glyph) = glyphs_iter.next() {
-            let copy_fun: fn(&mut TextureBuffer, &[u8], usize, (isize, isize));
+            let copy_fun: fn(&mut TextureBuffer, &[u8], usize, (usize, usize));
             let (stride, glyph_buffer) = match &glyph.buffer {
                 BitmapBuffer::Rgb(glyph_buffer) => {
                     copy_fun = TextureBuffer::write_rgb_at;
@@ -112,8 +112,8 @@ impl Rasterizer {
 
             if !glyph_buffer.is_empty() {
                 // Glyph position inside the buffer.
-                let y = (anchor_y + ascent) as isize - glyph.top as isize;
-                let x = ((anchor_x + offset) as i32 + glyph.left) as isize;
+                let y = anchor_y + ascent - glyph.top as usize;
+                let x = ((anchor_x + offset) as i32 + glyph.left) as usize;
 
                 // Copy the rasterized glyph to the output buffer.
                 let row_width = glyph.width as usize * stride;
